+++ HISTORY.md
@@ -6,6 +6,7 @@ Version 2.4.0 - 2013-10-XX
 --------------------------
 * Ruby 2.0 compatibility (copiousfreetime/stickler/#28
 * Implement latest-version command (copiousfreetime/stickler/#29)
+* Extract Server functionality to Middleware (copiousfreetime/stickler/#33)
 * Updated dependencies to latest versions
 
 Version 2.3.0 - 2013-03-09